requirements.txt
python-telegram-bot==20.7
Flask==2.3.3
openai==1.23.2